fixing conn disconnect handling from GnssLocationProvider
The "type" needs to maintain as signed, in case some value
may come across as negative, even though the storage size
may vary, from 8-bit to 32-bit, for example.

Bug: 65164961

CRs-Fixed: 2106350

+++ gnss/XtraSystemStatusObserver.cpp
@@ -61,11 +61,11 @@ bool XtraSystemStatusObserver::updateLockStatus(uint32_t lock) {
     return ( sendEvent(ss) );
 }
 
-bool XtraSystemStatusObserver::updateConnectionStatus(bool connected, uint32_t type) {
+bool XtraSystemStatusObserver::updateConnectionStatus(bool connected, int32_t type) {
     stringstream ss;
     ss <<  "connection";
     ss << " " << (connected ? "1" : "0");
-    ss << " " << (int)type;
+    ss << " " << type;
     ss << "\n"; // append seperator
     return ( sendEvent(ss) );
 }
